Scroll to focused code
Scrolling is controlled by a new parameter to RevealCodeFocus and is off
by default, to avoid any possible compatibility issues, although it
would probably be more user-friendly in the long run to default to on.
(Using an options object would probably be also be better than a boolean
parameter.)

The implementation pulls "classList.add('focus')" into a helper function
and adds tracking of offset of the focused lines of code.